Rename MenuItems to MenuList
