+++ packages/docs/components/molecules/Menu/burger/app.twig
@@ -11,7 +11,7 @@
       data_component: 'MenuBtn',
     }
   } %}
-  <div data-component="MenuItems"
+  <div data-component="MenuList"
     class="z-10 fixed top-0 left-0 h-full w-1/3 py-4 transform -translate-x-full pointer-events-none">
     <div class="absolute inset-0 bg-blue-200"></div>
     <ul class="w-full">
@@ -21,7 +21,7 @@
           href="#">
           Subitem #1
         </a>
-        <div data-component="MenuItems"
+        <div data-component="MenuList"
           class="z-under absolute top-0 left-full w-full h-full py-4 transform -translate-x-full pointer-events-none">
           <div class="absolute inset-0 bg-green-200"></div>
           <ul>
@@ -31,7 +31,7 @@
                 class="relative block p-4 hover:bg-blue-300 focus:ring-2 ring-inset">
                 Subitem #1.1
               </a>
-              <div data-component="MenuItems"
+              <div data-component="MenuList"
                 class="z-under absolute top-0 left-full w-full h-full py-4 bg-red-300 transform -translate-x-full pointer-events-none">
                 <ul>
                   <li>
@@ -100,7 +100,7 @@
           href="#">
           Subitem #2
         </a>
-        <div data-component="MenuItems"
+        <div data-component="MenuList"
           class="z-under absolute top-0 left-full w-full h-full py-4 transform -translate-x-full pointer-events-none">
           <div class="absolute inset-0 bg-green-200"></div>
           <ul>
@@ -110,7 +110,7 @@
                 class="relative block p-4 hover:bg-blue-300 focus:ring-2 ring-inset">
                 Subitem #2.1
               </a>
-              <div data-component="MenuItems"
+              <div data-component="MenuList"
                 class="z-under absolute top-0 left-full w-full h-full py-4 bg-red-300 transform -translate-x-full pointer-events-none">
                 <ul>
                   <li>
@@ -155,7 +155,7 @@
           href="#">
           Subitem #3
         </a>
-        <div data-component="MenuItems"
+        <div data-component="MenuList"
           class="z-under absolute top-0 left-full w-full h-full py-4 transform -translate-x-full pointer-events-none">
           <div class="absolute inset-0 bg-green-200"></div>
           <ul>
@@ -165,7 +165,7 @@
                 class="relative block p-4 hover:bg-blue-300 focus:ring-2 ring-inset">
                 Subitem #3.1
               </a>
-              <div data-component="MenuItems"
+              <div data-component="MenuList"
                 class="z-under absolute top-0 left-full w-full h-full py-4 bg-red-300 transform -translate-x-full pointer-events-none">
                 <ul>
                   <li>
