+++ packages/docs/components/molecules/Menu/dropdown/app.twig
@@ -11,7 +11,7 @@
       data_component: 'MenuBtn',
     }
   } %}
-  <div data-component="MenuItems"
+  <div data-component="MenuList"
     data-option-enter-to="is-open"
     data-option-enter-active="transition duration-200 ease-out-expo"
     data-option-leave-active="transition duration-200 ease-out-expo"
