+++ packages/docs/components/molecules/Menu/mega-menu/app.twig
@@ -11,7 +11,7 @@
         <a data-component="MenuBtn" class="block p-4 text-center bg-gray-200 hover:bg-gray-300" href="#">
           Item #{{ item }}
         </a>
-        <div data-component="MenuItems"
+        <div data-component="MenuList"
           data-option-enter-active="transition duration-200 ease-out-expo"
           data-option-leave-active="transition duration-200 ease-out-expo"
           data-option-leave-to="transform scale-y-95 opacity-0 pointer-events-none"
@@ -23,7 +23,7 @@
                 <a data-component="MenuBtn" class="block p-4 bg-gray-200 hover:bg-gray-300" href="#">
                   SubItem #{{ item }}.{{ subitem }}
                 </a>
-                <div data-component="MenuItems"
+                <div data-component="MenuList"
                   data-option-enter-active="transition duration-200 ease-out-expo"
                   data-option-leave-active="transition duration-200 ease-out-expo"
                   data-option-leave-to="opacity-0 pointer-events-none"
