+++ packages/ui/molecules/Menu/index.js
@@ -1,3 +1,3 @@
 export { default as Menu } from './Menu.js';
 export { default as MenuBtn } from './MenuBtn.js';
-export { default as MenuItems } from './MenuItems.js';
+export { default as MenuList } from './MenuList.js';
